add encoding tests
